ref(replay): Extract `overwriteRecordDroppedEvent` into fn
